refactor(threads): claudeHistoryLoader 外移 ask-user / shadow-recovery / final-timing
- 1962→1077 行：claudeAskUserQuestion 372 + claudeShadowRecovery 407 + claudeAssistantFinalTiming 111
- parseClaudeHistoryMessagesWithShadowRecovery 等经主文件 re-export 保持外部锁面
- i18n 依赖随 ask-user 函数外迁
- baseline 重生成 + 新文件登记豁免
验证：tsc 零 error；claudeHistoryLoader.test.ts 51 passed；gate 绿
