+++ docs/architecture/large-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T16:02:02.536Z",
+  "generatedAt": "2026-08-30T16:15:46.691Z",
   "scope": "fail",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
