+++ src/features/threads/loaders/claudeHistoryLoader.ts
@@ -16,6 +16,26 @@ import {
   areEquivalentReasoningTexts,
   compactComparableConversationText,
 } from "../assembly/conversationNormalization";
+import { parseClaudeHistoryMessagesWithShadowRecovery } from "./claudeShadowRecovery";
+export {
+  parseClaudeHistoryMessagesWithShadowRecovery,
+  recoverClaudeInterruptedAssistantFromShadow,
+} from "./claudeShadowRecovery";
+import {
+  extractClaudeAssistantFinalFlag,
+  hydrateClaudeAssistantFinalTiming,
+  markClaudeAssistantFinalMessages,
+} from "./claudeAssistantFinalTiming";
+import {
+  buildRequestUserInputSubmittedPayload,
+  isClaudeAskUserQuestionToolName,
+  isNativeClaudeAskUserQuestionToolName,
+  normalizeClaudeToolName,
+  parseAskUserQuestionAnswerText,
+  parseAskUserQuestionTemplates,
+  type AskUserQuestionAnswerParseResult,
+  type AskUserQuestionTemplate,
+} from "./claudeAskUserQuestion";
 import {
   buildClaudeControlEventItem,
   classifyClaudeControlPlaneMessage,
@@ -24,17 +44,12 @@ import {
   getClaudeHistoryMessageRole,
 } from "./claudeControlPlaneClassifier";
 import {
-  asBooleanFlag,
   asFiniteNonNegativeNumber,
   asRecord,
   buildUnifiedDiff,
   getFirstStringFieldFromRecords,
   parseHistoryTimestampMs,
-  parseJsonRecordFromText,
 } from "./claudeHistoryPrimitives";
-import { findLiveAssistantShadowTranscriptForRestore } from "../utils/liveAssistantShadowTranscript";
-import { noteThreadRecoverySourceObserved } from "../utils/streamLatencyDiagnostics";
-import { isCliInjectedAgentTaskNotificationText } from "../../engine-task-output/contracts/agentTaskNotification";
 import type { HistoryLoadingProgressListener } from "../utils/historyLoadingProgress";
 import { runNativeHistoryFetchAndParse } from "../utils/runNativeHistoryOpenStages";
 import { asString } from "./historyLoaderUtils";
@@ -51,59 +66,6 @@ type ClaudeHistoryLoaderOptions = {
 };
 
 export const CLAUDE_UI_HISTORY_WINDOW = 80;
-
-type MessageItemWithTurn = Extract<ConversationItem, { kind: "message" }> & {
-  role: "assistant";
-  turnId?: string | null;
-};
-
-type AskUserQuestionOption = {
-  label: string;
-  description: string;
-};
-
-type AskUserQuestionTemplate = {
-  id: string;
-  header: string;
-  question: string;
-  isOther: boolean;
-  multiSelect: boolean;
-  options?: AskUserQuestionOption[];
-};
-
-type AskUserQuestionAnswer = {
-  selectedOptions: string[];
-  note: string;
-};
-
-type AskUserQuestionAnswerParseResult = {
-  rawSelectionText: string;
-  answers: AskUserQuestionAnswer[];
-  answersByQuestionId: Record<string, AskUserQuestionAnswer>;
-};
-
-const ASK_USER_QUESTION_DISMISSED_TEXT_REGEX =
-  /^The user dismissed the question without selecting an option\.?$/i;
-const ASK_USER_QUESTION_SKIPPED_TEXT_REGEX =
-  /^The user skipped this AskUserQuestion without selecting an option\.\s*Do not ask the same question again;\s*continue the original task using the available context and reasonable assumptions\.?$/i;
-const ASK_USER_QUESTION_PARTIAL_SKIP_TEXT_REGEX =
-  /^The user answered the AskUserQuestion[:：]\s*([\s\S]*?)[。.]?\s*The user skipped \d+ remaining question\(s\) without selecting an option\.\s*Do not ask the skipped question\(s\) again;\s*continue the original task using the available context and reasonable assumptions\.?$/i;
-const ASK_USER_QUESTION_RESULT_BASE64_REGEX =
-  /\bAskUserQuestionResultBase64:([A-Za-z0-9+/=]+)/;
-
-type RequestUserInputSubmittedPayload = {
-  schema: "requestUserInputSubmitted/v1";
-  submittedAt: number;
-  questions: Array<{
-    id: string;
-    header: string;
-    question: string;
-    options?: AskUserQuestionOption[];
-    selectedOptions: string[];
-    note: string;
-  }>;
-};
-
 
 function isReasoningSnapshotDuplicate(previous: string, incoming: string) {
   return areEquivalentReasoningTexts(previous, incoming);
@@ -336,333 +298,6 @@ function getClaudeToolOutputRecord(message: Record<string, unknown>) {
   return null;
 }
 
-function parseAskUserQuestionTemplates(
-  toolInput: Record<string, unknown> | null,
-): AskUserQuestionTemplate[] {
-  if (!toolInput) {
-    return [];
-  }
-  const hasSingleQuestionShape =
-    "question" in toolInput ||
-    "prompt" in toolInput ||
-    "header" in toolInput ||
-    "title" in toolInput ||
-    "options" in toolInput;
-  const rawQuestions = Array.isArray(toolInput.questions)
-    ? toolInput.questions
-    : hasSingleQuestionShape
-      ? [toolInput]
-      : [];
-  const templates: AskUserQuestionTemplate[] = [];
-  rawQuestions.forEach((entry, index) => {
-    const question = asRecord(entry);
-    if (!question) {
-      return;
-    }
-    const id = asString(question.id ?? `q-${index}`).trim() || `q-${index}`;
-    const header = asString(question.header ?? question.title ?? "").trim();
-    const questionText = asString(
-      question.question ?? question.prompt ?? "",
-    ).trim();
-    const isOther =
-      question.isOther === undefined && question.is_other === undefined
-        ? true
-        : Boolean(question.isOther ?? question.is_other);
-    const multiSelect = Boolean(question.multiSelect ?? question.multi_select);
-    const rawOptions = Array.isArray(question.options) ? question.options : [];
-    const options = rawOptions
-      .map((rawOption) => {
-        const option = asRecord(rawOption);
-        if (!option) {
-          return null;
-        }
-        const label = asString(option.label ?? "").trim();
-        const description = asString(option.description ?? "").trim();
-        if (!label && !description) {
-          return null;
-        }
-        return { label, description };
-      })
-      .filter((option): option is AskUserQuestionOption => option !== null);
-    if (!questionText && options.length === 0) {
-      return;
-    }
-    templates.push({
-      id,
-      header,
-      question: questionText,
-      isOther,
-      multiSelect,
-      options: options.length > 0 ? options : undefined,
-    });
-  });
-  return templates;
-}
-
-function parseAskUserAnswerParts(raw: string): AskUserQuestionAnswer {
-  const segments = raw
-    .split(/[,，、]/)
-    .map((part) => part.trim())
-    .filter(Boolean);
-  const selectedOptions: string[] = [];
-  let note = "";
-  for (const segment of segments) {
-    if (/^user_note\s*:/i.test(segment)) {
-      const parsedNote = segment.replace(/^user_note\s*:/i, "").trim();
-      if (parsedNote) {
-        note = parsedNote;
-      }
-      continue;
-    }
-    selectedOptions.push(segment);
-  }
-  return { selectedOptions, note };
-}
-
-function decodeBase64JsonRecord(
-  base64Value: string,
-): Record<string, unknown> | null {
-  try {
-    const binary = globalThis.atob(base64Value);
-    const bytes = Uint8Array.from(binary, (char) => char.charCodeAt(0));
-    const jsonText = new TextDecoder().decode(bytes);
-    return parseJsonRecordFromText(jsonText);
-  } catch {
-    return null;
-  }
-}
-
-function parseStructuredAnswerResult(
-  text: string,
-  templates: AskUserQuestionTemplate[],
-): AskUserQuestionAnswerParseResult | null {
-  const markerMatch = text.match(ASK_USER_QUESTION_RESULT_BASE64_REGEX);
-  if (!markerMatch) {
-    return null;
-  }
-  const payload = decodeBase64JsonRecord(asString(markerMatch[1]));
-  const answersRecord = asRecord(payload?.answers);
-  if (!answersRecord) {
-    return null;
-  }
-  const templateIds = new Set(templates.map((template) => template.id));
-  const answersByQuestionId: Record<string, AskUserQuestionAnswer> = {};
-  const displayParts: string[] = [];
-  Object.entries(answersRecord).forEach(([questionId, rawAnswers]) => {
-    if (!templateIds.has(questionId) || !Array.isArray(rawAnswers)) {
-      return;
-    }
-    const answerValues = rawAnswers
-      .map((value) => asString(value).trim())
-      .filter(Boolean);
-    if (answerValues.length === 0) {
-      return;
-    }
-    const answerText = answerValues.join(", ");
-    answersByQuestionId[questionId] = parseAskUserAnswerParts(answerText);
-    displayParts.push(answerText);
-  });
-  if (Object.keys(answersByQuestionId).length === 0) {
-    return null;
-  }
-  return {
-    rawSelectionText: displayParts.join("; "),
-    answers: [],
-    answersByQuestionId,
-  };
-}
-
-function parseAskUserAnswerSegments(
-  rawSelectionText: string,
-  templates: AskUserQuestionTemplate[],
-) {
-  const templateIds = new Set(templates.map((template) => template.id));
-  const baseSegments = rawSelectionText
-    .split(/[;；]/)
-    .map((segment) => segment.trim())
-    .filter(Boolean);
-  const answersByQuestionId: Record<string, AskUserQuestionAnswer> = {};
-  const positionalAnswerSegments: string[] = [];
-  for (const segment of baseSegments) {
-    const keyedMatch = segment.match(/^([A-Za-z0-9_.:-]+)\s*=\s*([\s\S]*)$/);
-    if (keyedMatch) {
-      const questionId = asString(keyedMatch[1]).trim();
-      const answerText = asString(keyedMatch[2]).trim();
-      if (templateIds.has(questionId)) {
-        answersByQuestionId[questionId] = parseAskUserAnswerParts(answerText);
-        continue;
-      }
-    }
-    positionalAnswerSegments.push(segment);
-  }
-  return {
-    answersByQuestionId,
-    positionalAnswerSegments,
-    displaySelectionText: baseSegments
-      .map((segment) => {
-        const keyedMatch = segment.match(
-          /^([A-Za-z0-9_.:-]+)\s*=\s*([\s\S]*)$/,
-        );
-        if (!keyedMatch) {
-          return segment;
-        }
-        const questionId = asString(keyedMatch[1]).trim();
-        return templateIds.has(questionId)
-          ? asString(keyedMatch[2]).trim()
-          : segment;
-      })
-      .filter(Boolean)
-      .join("; "),
-  };
-}
-
-function parseAskUserQuestionAnswerText(
-  text: string,
-  templates: AskUserQuestionTemplate[],
-): AskUserQuestionAnswerParseResult | null {
-  const trimmed = text.trim();
-  const questionCount = templates.length;
-  if (!trimmed) {
-    return null;
-  }
-
-  if (
-    ASK_USER_QUESTION_DISMISSED_TEXT_REGEX.test(trimmed) ||
-    ASK_USER_QUESTION_SKIPPED_TEXT_REGEX.test(trimmed)
-  ) {
-    return {
-      rawSelectionText: "",
-      answers: Array.from({ length: Math.max(questionCount, 1) }, () => ({
-        selectedOptions: [],
-        note: "",
-      })),
-      answersByQuestionId: {},
-    };
-  }
-
-  const structuredResult = parseStructuredAnswerResult(trimmed, templates);
-  if (structuredResult) {
-    return structuredResult;
-  }
-
-  const answeredMatch =
-    trimmed.match(ASK_USER_QUESTION_PARTIAL_SKIP_TEXT_REGEX) ??
-    trimmed.match(
-      /^The user answered the AskUserQuestion:\s*([\s\S]*?)(?:[。.]?\s*Please continue based on this selection\.?)$/i,
-    );
-  if (!answeredMatch) {
-    return null;
-  }
-  const rawSelectionText = (answeredMatch[1] ?? "").trim();
-  if (!rawSelectionText) {
-    return null;
-  }
-
-  const {
-    answersByQuestionId,
-    positionalAnswerSegments,
-    displaySelectionText,
-  } = parseAskUserAnswerSegments(rawSelectionText, templates);
-  if (
-    positionalAnswerSegments.length === 0 &&
-    Object.keys(answersByQuestionId).length === 0
-  ) {
-    return null;
-  }
-  if (Object.keys(answersByQuestionId).length > 0) {
-    return {
-      rawSelectionText: displaySelectionText,
-      answers: [],
-      answersByQuestionId,
-    };
-  }
-  if (questionCount <= 1) {
-    return {
-      rawSelectionText,
-      answers: [parseAskUserAnswerParts(rawSelectionText)],
-      answersByQuestionId,
-    };
-  }
-
-  const normalizedSegments = [...positionalAnswerSegments];
-  if (normalizedSegments.length > questionCount) {
-    const remaining = normalizedSegments
-      .splice(questionCount - 1)
-      .join("; ")
-      .trim();
-    normalizedSegments.push(remaining);
-  }
-  while (normalizedSegments.length < questionCount) {
-    normalizedSegments.push("");
-  }
-
-  return {
-    rawSelectionText,
-    answers: normalizedSegments.map((segment) =>
-      parseAskUserAnswerParts(segment),
-    ),
-    answersByQuestionId,
-  };
-}
-
-function buildRequestUserInputSubmittedPayload(
-  templates: AskUserQuestionTemplate[],
-  parseResult: AskUserQuestionAnswerParseResult,
-): RequestUserInputSubmittedPayload {
-  return {
-    schema: "requestUserInputSubmitted/v1",
-    submittedAt: Date.now(),
-    questions: templates.map((template, index) => ({
-      id: template.id || `q-${index}`,
-      header: template.header,
-      question: template.question,
-      options: template.options,
-      selectedOptions:
-        parseResult.answersByQuestionId[template.id]?.selectedOptions ??
-        parseResult.answers[index]?.selectedOptions ??
-        [],
-      note:
-        parseResult.answersByQuestionId[template.id]?.note ??
-        parseResult.answers[index]?.note ??
-        "",
-    })),
-  };
-}
-
-function normalizeClaudeToolName(toolName: string) {
-  return toolName.trim().toLowerCase();
-}
-
-/** Native AskUserQuestion or MCP bridge (`mcp__ccgui__AskUserQuestion`). */
-function isClaudeAskUserQuestionToolName(toolName: string) {
-  const normalized = normalizeClaudeToolName(toolName);
-  if (!normalized) {
-    return false;
-  }
-  if (
-    normalized === "askuserquestion" ||
-    normalized === "ask_user_question"
-  ) {
-    return true;
-  }
-  return (
-    normalized.includes("askuserquestion") ||
-    normalized.includes("ask_user_question")
-  );
-}
-
-/**
- * MCP bridge asks cannot be answered after history reopen: live request_id is a
- * hash of a synthetic tool_id, while transcript tool ids are CLI tool_use ids.
- * Only rehydrate interactive cards for native AskUserQuestion.
- */
-function isNativeClaudeAskUserQuestionToolName(toolName: string) {
-  const normalized = normalizeClaudeToolName(toolName);
-  return (
-    normalized === "askuserquestion" || normalized === "ask_user_question"
-  );
-}
-
 const CLAUDE_FILE_PATH_KEYS = [
   "file_path",
   "filePath",
@@ -817,118 +452,6 @@ function mergeReasoningSnapshot(
     kind: "reasoning",
     summary: normalizedText.slice(0, 100),
     content: normalizedText,
-  });
-}
-
-function extractClaudeAssistantFinalFlag(
-  message: Record<string, unknown>,
-): boolean | undefined {
-  const metadata =
-    message.metadata &&
-    typeof message.metadata === "object" &&
-    !Array.isArray(message.metadata)
-      ? (message.metadata as Record<string, unknown>)
-      : null;
-  const candidates: unknown[] = [
-    message.isFinal,
-    message.is_final,
-    message.final,
-    message.isFinalMessage,
-    message.is_final_message,
-  ];
-  if (metadata) {
-    candidates.push(
-      metadata.isFinal,
-      metadata.is_final,
-      metadata.final,
-      metadata.isFinalMessage,
-      metadata.is_final_message,
-    );
-  }
-  for (const candidate of candidates) {
-    const parsed = asBooleanFlag(candidate);
-    if (typeof parsed === "boolean") {
-      return parsed;
-    }
-  }
-  return undefined;
-}
-
-function markClaudeAssistantFinalMessages(items: ConversationItem[]) {
-  let lastAssistantIndexInTurn = -1;
-  let hasExplicitFinalAssistantInTurn = false;
-  const finalizeCurrentTurn = () => {
-    if (hasExplicitFinalAssistantInTurn || lastAssistantIndexInTurn < 0) {
-      return;
-    }
-    const lastAssistant = items[lastAssistantIndexInTurn];
-    if (
-      !lastAssistant ||
-      lastAssistant.kind !== "message" ||
-      lastAssistant.role !== "assistant"
-    ) {
-      return;
-    }
-    if (lastAssistant.isFinal === true) {
-      return;
-    }
-    items[lastAssistantIndexInTurn] = {
-      ...lastAssistant,
-      isFinal: true,
-    };
-  };
-
-  items.forEach((item, index) => {
-    if (item.kind === "message" && item.role === "user") {
-      finalizeCurrentTurn();
-      lastAssistantIndexInTurn = -1;
-      hasExplicitFinalAssistantInTurn = false;
-      return;
-    }
-    if (item.kind === "message" && item.role === "assistant") {
-      if (item.isFinal === true) {
-        hasExplicitFinalAssistantInTurn = true;
-      }
-      lastAssistantIndexInTurn = index;
-    }
-  });
-
-  finalizeCurrentTurn();
-}
-
-function hydrateClaudeAssistantFinalTiming(
-  items: ConversationItem[],
-  messageTimestampById: Map<string, number>,
-) {
-  let turnStartedAtMs: number | undefined;
-  items.forEach((item, index) => {
-    if (item.kind !== "message") {
-      return;
-    }
-    if (item.role === "user") {
-      turnStartedAtMs = messageTimestampById.get(item.id);
-      return;
-    }
-    if (item.isFinal !== true) {
-      return;
-    }
-    const completedAtMs = messageTimestampById.get(item.id);
-    const durationMs =
-      typeof completedAtMs === "number" && typeof turnStartedAtMs === "number"
-        ? Math.max(0, completedAtMs - turnStartedAtMs)
-        : undefined;
-    if (typeof completedAtMs !== "number" && typeof durationMs !== "number") {
-      return;
-    }
-    items[index] = {
-      ...item,
-      ...(typeof completedAtMs === "number"
-        ? { finalCompletedAt: completedAtMs }
-        : {}),
-      ...(typeof durationMs === "number"
-        ? { finalDurationMs: durationMs }
-        : {}),
-    };
   });
 }
 
@@ -1436,414 +959,6 @@ function extractPendingUserInputQueueFromClaudeItems(
   }
 
   return queue;
-}
-
-function isShadowRecoveryTurnBoundaryUser(item: ConversationItem) {
-  return (
-    item.kind === "message" &&
-    item.role === "user" &&
-    !isCliInjectedAgentTaskNotificationText(item.text)
-  );
-}
-
-function isEquivalentShadowAssistantText(existingText: string, shadowText: string) {
-  const normalizedExisting = existingText.trim();
-  const normalizedShadow = shadowText.trim();
-  if (!normalizedExisting || !normalizedShadow) {
-    return false;
-  }
-  return (
-    normalizedShadow === normalizedExisting ||
-    normalizedShadow.startsWith(normalizedExisting) ||
-    normalizedExisting.startsWith(normalizedShadow)
-  );
-}
-
-function findExistingAssistantEquivalentToShadow(
-  items: ConversationItem[],
-  shadowText: string,
-) {
-  for (let index = items.length - 1; index >= 0; index -= 1) {
-    const item = items[index];
-    if (item?.kind !== "message" || item.role !== "assistant") {
-      continue;
-    }
-    if (isEquivalentShadowAssistantText(item.text, shadowText)) {
-      return { index, item };
-    }
-  }
-  return null;
-}
-
-function findLastAssistantAfterLastUser(items: ConversationItem[]) {
-  let lastUserIndex = -1;
-  for (let index = 0; index < items.length; index += 1) {
-    const item = items[index];
-    if (item && isShadowRecoveryTurnBoundaryUser(item)) {
-      lastUserIndex = index;
-    }
-  }
-  if (lastUserIndex < 0) {
-    return null;
-  }
-  for (let index = items.length - 1; index > lastUserIndex; index -= 1) {
-    const item = items[index];
-    if (item?.kind === "message" && item.role === "assistant") {
-      return { index, item };
-    }
-  }
-  return null;
-}
-
-function isLikelySameTurnForRecovery(
-  item: MessageItemWithTurn,
-  expectedTurnId: string | null,
-  shadowTurnId: string | null,
-) {
-  if (!expectedTurnId || !item.turnId || !shadowTurnId) {
-    return true;
-  }
-  return item.turnId === expectedTurnId && item.turnId === shadowTurnId;
-}
-
-function buildRecoveredAssistantFromShadow(shadow: {
-  id: string;
-  text: string;
-  turnId: string | null;
-  itemId: string;
-}) {
-  return {
-    id: `claude-shadow-recovered-${shadow.itemId}`,
-    kind: "message" as const,
-    role: "assistant" as const,
-    text: shadow.text,
-    ...(shadow.turnId ? { turnId: shadow.turnId } : {}),
-    isFinal: false,
-    recoveredFromLiveShadow: true,
-    recoveryStatus: "interrupted" as const,
-    recoverySourceId: shadow.id,
-    engineSource: "claude" as const,
-  };
-}
-
-function appendRecoveredAssistantUnlessDuplicate({
-  items,
-  shadow,
-  expectedTurnId,
-  hasExplicitFinalAfterLastUser,
-}: {
-  items: ConversationItem[];
-  shadow: {
-    id: string;
-    text: string;
-    turnId: string | null;
-    itemId: string;
-  };
-  expectedTurnId: string | null;
-  hasExplicitFinalAfterLastUser: boolean;
-}) {
-  const equivalent = findExistingAssistantEquivalentToShadow(items, shadow.text);
-  if (!equivalent) {
-    return [...items, buildRecoveredAssistantFromShadow(shadow)];
-  }
-
-  const latest = equivalent.item as MessageItemWithTurn;
-  const normalizedLatestText = latest.text.trim();
-  if (latest.isFinal === true && hasExplicitFinalAfterLastUser) {
-    return items;
-  }
-  const sameTurn = isLikelySameTurnForRecovery(
-    latest,
-    expectedTurnId,
-    shadow.turnId,
-  );
-  const normalizedShadowText = shadow.text.trim();
-  if (
-    sameTurn &&
-    normalizedShadowText.startsWith(normalizedLatestText) &&
-    normalizedShadowText.length > normalizedLatestText.length
-  ) {
-    const merged = items.slice();
-    merged[equivalent.index] = {
-      ...latest,
-      text: shadow.text,
-      isFinal: false,
-      recoveredFromLiveShadow: true,
-      recoveryStatus: "interrupted" as const,
-      recoverySourceId: shadow.id,
-      engineSource: "claude" as const,
-      ...(shadow.turnId ? { turnId: shadow.turnId } : {}),
-    };
-    return merged;
-  }
-  return items;
-}
-
-function recoverClaudeAssistantFromShadowIfNeeded({
-  items,
-  shadow,
-  expectedTurnId,
-  hasExplicitFinalAfterLastUser,
-}: {
-  items: ConversationItem[];
-  shadow: {
-    id: string;
-    text: string;
-    turnId: string | null;
-    itemId: string;
-  };
-  expectedTurnId: string | null;
-  hasExplicitFinalAfterLastUser: boolean;
-}) {
-  const normalizedShadowText = shadow.text.trim();
-  if (normalizedShadowText.length === 0) {
-    return items;
-  }
-
-  const trailingAssistant = findLastAssistantAfterLastUser(items);
-  if (!trailingAssistant) {
-    return appendRecoveredAssistantUnlessDuplicate({
-      items,
-      shadow,
-      expectedTurnId,
-      hasExplicitFinalAfterLastUser,
-    });
-  }
-
-  const latest = trailingAssistant.item as MessageItemWithTurn;
-  const normalizedLatestText = latest.text.trim();
-  if (!normalizedLatestText) {
-    return appendRecoveredAssistantUnlessDuplicate({
-      items,
-      shadow,
-      expectedTurnId,
-      hasExplicitFinalAfterLastUser,
-    });
-  }
-
-  const sameTurn = isLikelySameTurnForRecovery(
-    latest,
-    expectedTurnId,
-    shadow.turnId,
-  );
-
-  if (latest.isFinal === true && hasExplicitFinalAfterLastUser) {
-    return items;
-  }
-
-  const isPrefixMatch = normalizedShadowText.startsWith(normalizedLatestText);
-
-  if (sameTurn && isPrefixMatch) {
-    if (normalizedShadowText.length > normalizedLatestText.length) {
-      const merged = items.slice();
-      merged[trailingAssistant.index] = {
-        ...latest,
-        text: shadow.text,
-        isFinal: false,
-        recoveredFromLiveShadow: true,
-        recoveryStatus: "interrupted" as const,
-        recoverySourceId: shadow.id,
-        engineSource: "claude" as const,
-        ...(shadow.turnId ? { turnId: shadow.turnId } : {}),
-      };
-      return merged;
-    }
-    return items;
-  }
-
-  return items;
-}
-
-function deriveLatestTurnIdFromItems(items: ConversationItem[]) {
-  for (let index = items.length - 1; index >= 0; index -= 1) {
-    const item = items[index];
-    if (!item) {
-      continue;
-    }
-    const turnId = asString((item as { turnId?: string | null }).turnId).trim();
-    if (turnId) {
-      return turnId;
-    }
-  }
-  return null;
-}
-
-function deriveLatestTurnIdFromRawMessages(messagesData: unknown) {
-  const messages = Array.isArray(messagesData) ? messagesData : [];
-  for (let index = messages.length - 1; index >= 0; index -= 1) {
-    const message = asRecord(messages[index]);
-    if (!message) {
-      continue;
-    }
-    const turnId = extractTurnIdFromRawMessage(message);
-    if (turnId) {
-      return turnId;
-    }
-  }
-  return null;
-}
-
-function resolveExpectedShadowTurnId(
-  items: ConversationItem[],
-  messagesData: unknown,
-): string | null {
-  const rawTurnId = deriveLatestTurnIdFromRawMessages(messagesData);
-  if (rawTurnId) {
-    return rawTurnId;
-  }
-  return deriveLatestTurnIdFromItems(items);
-}
-
-function hasExplicitFinalAssistantAfterLastUser(messagesData: unknown) {
-  const messages = Array.isArray(messagesData) ? messagesData : [];
-  let lastUserIndex = -1;
-  for (let index = 0; index < messages.length; index += 1) {
-    const message = asRecord(messages[index]);
-    if (!message) {
-      continue;
-    }
-    const kind = asString(message.kind ?? "");
-    if (kind !== "message") {
-      continue;
-    }
-    const role = asString(message.role ?? "") === "user" ? "user" : "assistant";
-    if (
-      role === "user" &&
-      !isCliInjectedAgentTaskNotificationText(asString(message.text ?? ""))
-    ) {
-      lastUserIndex = index;
-    }
-  }
-  if (lastUserIndex < 0) {
-    return false;
-  }
-
-  for (let index = lastUserIndex + 1; index < messages.length; index += 1) {
-    const message = asRecord(messages[index]);
-    if (!message) {
-      continue;
-    }
-    const kind = asString(message.kind ?? "");
-    if (kind !== "message") {
-      continue;
-    }
-    const role = asString(message.role ?? "") === "user" ? "user" : "assistant";
-    if (role !== "assistant") {
-      continue;
-    }
-    if (extractClaudeAssistantFinalFlag(message) === true) {
-      return true;
-    }
-  }
-  return false;
-}
-
-export function recoverClaudeInterruptedAssistantFromShadow({
-  items,
-  workspaceId,
-  threadId,
-  sessionId,
-  expectedTurnId,
-  hasExplicitFinalAfterLastUser,
-}: {
-  items: ConversationItem[];
-  workspaceId: string;
-  threadId: string;
-  sessionId: string;
-  expectedTurnId?: string | null;
-  hasExplicitFinalAfterLastUser?: boolean;
-}) {
-  const resolvedExpectedTurnId = expectedTurnId?.trim() || null;
-  const lookupTurnId =
-    resolvedExpectedTurnId || deriveLatestTurnIdFromItems(items);
-  const unsettledShadow = findLiveAssistantShadowTranscriptForRestore({
-    workspaceId,
-    threadId,
-    sessionId,
-    requireUnsettled: true,
-    expectedTurnId: lookupTurnId,
-  });
-
-  const shadow =
-    unsettledShadow ??
-    findLiveAssistantShadowTranscriptForRestore({
-      workspaceId,
-      threadId,
-      sessionId,
-      requireUnsettled: false,
-      ...(lookupTurnId ? { expectedTurnId: lookupTurnId } : {}),
-    });
-
-  if (!shadow) {
-    return items;
-  }
-
-  const recoveredItems = recoverClaudeAssistantFromShadowIfNeeded({
-    items,
-    shadow: {
-      id: shadow.id,
-      text: shadow.text,
-      turnId: shadow.turnId,
-      itemId: shadow.itemId,
-    },
-    expectedTurnId: lookupTurnId,
-    hasExplicitFinalAfterLastUser: hasExplicitFinalAfterLastUser === true,
-  });
-  const recoveredAssistant = recoveredItems.find(
-    (item): item is Extract<ConversationItem, { kind: "message" }> =>
-      item.kind === "message" &&
-      item.role === "assistant" &&
-      item.recoveredFromLiveShadow === true &&
-      item.recoverySourceId === shadow.id,
-  );
-  if (recoveredAssistant) {
-    noteThreadRecoverySourceObserved(threadId, {
-      source: "live-shadow",
-      sourceId: shadow.id,
-      itemId: recoveredAssistant.id,
-      textLength: recoveredAssistant.text.length,
-    });
-  }
-  return recoveredItems;
-}
-
-function normalizeShadowRecoverySessionId(
-  threadId: string,
-  sessionId?: string | null,
-) {
-  const candidate = sessionId?.trim();
-  if (candidate) {
-    return candidate;
-  }
-  const threadPrefix = threadId.startsWith("claude:")
-    ? threadId.slice("claude:".length)
-    : threadId;
-  return threadPrefix.trim();
-}
-
-export function parseClaudeHistoryMessagesWithShadowRecovery({
-  messagesData,
-  workspaceId,
-  threadId,
-  sessionId,
-  workspacePath,
-}: {
-  messagesData: unknown;
-  workspaceId: string;
-  threadId: string;
-  sessionId?: string | null;
-  workspacePath?: string | null;
-}): ConversationItem[] {
-  const parsedItems = parseClaudeHistoryMessages(messagesData, workspacePath);
-  return recoverClaudeInterruptedAssistantFromShadow({
-    items: parsedItems,
-    workspaceId,
-    threadId,
-    sessionId: normalizeShadowRecoverySessionId(threadId, sessionId),
-    expectedTurnId: resolveExpectedShadowTurnId(parsedItems, messagesData),
-    hasExplicitFinalAfterLastUser:
-      hasExplicitFinalAssistantAfterLastUser(messagesData),
-  });
 }
 
 export function extractClaudeHistoryTokenUsage(
